Rename and unexport methods for clarity, fix error messaging
Hide methods (`formUrl` and `parseJson`) that will only be used
internally. Move tests for those methods into the `weather` package

Rename methods (`formURl` to `formAPIUrl`) to be more descriptive.

The method that returns a forecast is now named `ForecastByCity`, and
returns the forecast instead of requiring the forecast to be fetched.